// data/github/current/ui/packages/copilot-for-business/types.ts
export type SeatBreakdown = {
  seats_assigned: number
  seats_billed: number
  seats_pending: number
  description: string
}

export type CopilotForBusinessTrial = {
  started: boolean
  ended: boolean
  has_trial: boolean
  upgradable: boolean
  cancelable: boolean
  days_left: number
  started_at: string
  ends_at: string
  trial_length: number
  active: boolean
  expired: boolean
  pending: boolean
  copilot_plan: string
  belongs_to_trial_business_account: boolean
}
// ...
export type PlanText = 'Business' | 'Enterprise'